do unit test 1 - uocso
